Refactor smw helper functions into own module
