+++ src/updaters/wikiVendor.js
@@ -1,9 +1,9 @@
-const querystring = require('querystring')
 const fetch = require('node-fetch')
 
-const BASE_API_URL = 'https://wiki.guildwars2.com/api.php?'
-
-let currencies = {}
+const smw = require('./smw')
+const queryApi = smw.queryApi
+const getIdForName = smw.getIdForName
+const getCurrencyForName = smw.getCurrencyForName
 
 module.exports.getVendorsForItem = getVendorsForItem
 module.exports.getVendor = getVendor
@@ -27,8 +27,6 @@ async function getVendorsForItem(nameOrId) {
 async function getVendor(vendorName) {
   console.log(`Processing vendor ${vendorName}`)
 
-  currencies = await getCurrencies()
-
   const offeredItems = await queryApi(`[[Has vendor::${vendorName}]]|?Sells item|?Sells item.Has game id|?Has item quantity|?Has item cost|?Has availability`)
   if (offeredItems.length === 0) {
     console.error(`Vendor ${vendorName} not found.`)
@@ -40,26 +38,6 @@ async function getVendor(vendorName) {
     locations: await getLocations(vendorName),
     purchase_options: (await Promise.all(offeredItems.map(formatOfferedItem))).filter(Boolean)
   }
-}
-
-async function queryApi (query, offset = 0) {
-  const parameters = {
-    action: 'ask',
-    format: 'json',
-    query: query + `|limit=500|offset=${offset}`,
-  }
-
-  const url = BASE_API_URL + querystring.stringify(parameters)
-  const response = await fetch(url).then(x => x.json())
-
-  let results = Object.values(response.query.results)
-  let moreResults = []
-
-  if (response['query-continue-offset']) {
-    moreResults = await queryApi(query, response['query-continue-offset'])
-  }
-
-  return results.concat(moreResults)
 }
 
 async function formatOfferedItem (item) {
@@ -90,7 +68,7 @@ async function formatCost (cost) {
 
   const name = cost['Has item currency'].item[0]
 
-  const {currency, multiplier} = getCurrency(name)
+  const {currency, multiplier} = await getCurrencyForName(name)
   if (currency) {
     result.type = 'Currency'
     result.id = currency.id
@@ -99,10 +77,7 @@ async function formatCost (cost) {
     result.id = await getIdForName(name)
   }
 
-  result.count = Number(cost['Has item value'].item[0])
-  if (multiplier) {
-    result.count = result.count * multiplier
-  }
+  result.count = multiplier * Number(cost['Has item value'].item[0])
 
   if (result.type === undefined || result.id === undefined || result.count === undefined) {
     console.error('Cost misses some property:', result)
@@ -113,64 +88,6 @@ async function formatCost (cost) {
   result.count = result.count || null
 
   return result
-}
-
-function getCurrency(name) {
-  let multiplier = undefined
-  switch (name) {
-    case 'Ascended Shard of Glory':
-      name = 'Ascended Shards of Glory'
-      break
-    case 'Laurels':
-      name = 'Laurel'
-      break
-    case 'Gold':
-      name = 'Coin'
-      multiplier = 10000
-      break
-    case 'Silver':
-      name = 'Coin'
-      multiplier = 100
-      break
-    case 'Copper':
-      name = 'Coin'
-      break
-    default:
-      break
-  }
-
-  const currency = currencies.find((x) => x.name.toLowerCase() === name.toLowerCase())
-
-  return {currency, multiplier}
-}
-
-async function getIdForName(name) {
-  if (name === 'Ectoplasm') { // instead of 'Glob of Ectoplasm'
-    return 19721
-  }
-
-  // agony infusions have their "+" escaped (=removed) in the vendor property, but can only be queried for with the full name...
-  if (/^(\w+ )?[0-9]+ (Agony|Simple) Infusion$/.exec(name)) {
-    const index = /[0-9]+ (Agony|Simple) Infusion$/.exec(name).index
-    name = name.substring(0, index) + "+" + name.substring(index)
-  }
-
-  const results = await queryApi(`[[Has context::Item]][[Has canonical name::${name}]]|?Has game id`)
-
-  if (results.length === 0) {
-    console.error(`Found no id for '${name}'. Please fix this manually (look out for 'id: null').`)
-    return null
-  }
-
-  if (results.length > 1) {
-    console.warn(`Name '${name}' is ambiguous. Found ids ${results.map((x) => x.printouts['Has game id'][0])}. Using the first one.`)
-  }
-
-  return results[0].printouts['Has game id'][0]
-}
-
-async function getCurrencies() {
-  return await fetch('https://api.guildwars2.com/v2/currencies?lang=en&ids=all').then(x => x.json())
 }
 
 async function getLocations(vendorName) {
